requirements.txt
streamlit
pandas
mysql-connector-python
joblib
fastapi
uvicorn
scikit-learn